moved enumerations in azure::core::lease
